Improved algorithms
I improved my previous subsolar point calculation algorithms. Increases the accuracy and efficiency of the calculations. It also uses more efficient algorithms for the equation of time, the mean anomaly, the equation of center, and the geometric mean longitude. The bottom line is that the real-time position of the Sun is now even more accurate.